move vault from initializer to a library
